Laravel Event with ShouldBroadcast and Broadcasting
In this part we demonstrate how to broadcast a message under WebSocket. Right now we can only determine whom should be sent a message by the message itself. The one actually already received that message, but he or she found that it does not belongs to him or her. We should build an authentication that could direct locate a specific user, by tne Event name or a unique Channel.